Adding ts types from core to client output.
In order for this to work nicely, a "typescript" feature was added to
core which adds `wasm-bindgen` and `typescript-definitions`, which are
used to annotate the types for the `index.d.ts` generated in `e2e-client`.

Another change made here was to add a rust-toolchain file to the client,
which now requires nightly to be able to do the ts output magic. Along
with this, I noticed that I didn't see build failures when running
`cargo build` directly on the client, but did when running the wasm-pack
build via npm. This was because wasm-pack switches the target off of
whatever the default is over to wasm32 instead. To cut down on
confusion, I configured cargo (in that one crate) to default to wasm32.