Update e2e-core to use juniper/juniper_from_schema.
This removes a couple types that are now being generated by the various
macros.

`Message`, for example, is an **input type** which is managed by
`juniper_from_schema`.

Also interesting to note, we had been operating with the assumption of
sharing the "shape" of `Message { author: string, text: string }`
between the create call, and as a field in the `ChatLogEntry` type, but
graphql puts a hard fence between types that can be returned as fields,
and types used as inputs.

As a result of this, the graphql view of `ChatLogEntry` is *flattened*
instead of just having a `Message` as a field in it.