Redo the actix storages so they can be shared with juniper Context.
Turns out most of the types we build for juniper need to be in the same
crate as where the storage is, since the types are all bound by the type
of the Context struct (which exposes the storage to them).

e2e-core is empty for now. Everything moved over to e2e-server.

Also, moved the storaged into a "data" module.

One notable struggle I had in this set of changes were regarding the RNG
for the name generator. ThreadRng cannot cross threads, and is difficult
to share from Data to Context. To sidestep the issue, I switched to
using OsRng to effectively "outsource" the generator to the system. This
should be fine for us.

At this point, the graphql engine is wired up, and able to evaluate
requests. Next steps are to wire up the types in the schema module to
the storages via the context they receive in their field resolver
methods.