Wiring in the graphql layer to leverage storages.
There's some repetition in lock management that we can refactor to
reduce. Looks a little messy right now.

Still, this rev brings us to "feature complete" status for the backend,
unblocking the frontend work.

Usernames are generating, and announcing the logins as they do. Messages
are available for listing, and we can push new messages into the storage
too. That's the whole thing.