Generate client code from protobuf.
Initial attempts to generate the grpc client seemed to produce only part
of it.
The new gen-client.sh script uses both --js_out and --grpc-web_out flags
(yes, both) to produce the entire thing. Looks like the binary output
from the js exporter is needed for the core transport code, whereas the
grpc-web export has the client that wraps it.

This rev gives us the client as an npm package, and also adds it as a
dependency in the web-frontend project so as to try and verify the
symbols exported are able to load correctly with the base CRA loaders.